V3.0 feature icon frame

+++ DMHelper/src/audiotracksyrinscapeonline.h
@@ -20,7 +20,7 @@ public slots:
     virtual void play() override;
     virtual void stop() override;
     virtual void setMute(bool mute) override;
-    virtual void setVolume(int volume) override;
+    virtual void setVolume(float volume) override;
     virtual void setRepeat(bool repeat) override;
 
 protected slots:
